+++ components/usergroup/latestTech/divBadge.tsx
@@ -1,0 +1,5 @@
+export const DIVBadge = (props: DivBadgeProps) => {
+  console.log(props);
+
+  return <div></div>;
+};
